+++ app/controllers/SampleController.go
@@ -6,10 +6,10 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/kosatnkn/catalyst/app/container"
-	"github.com/kosatnkn/catalyst/app/transport/request"
-	"github.com/kosatnkn/catalyst/app/transport/request/unpackers"
-	"github.com/kosatnkn/catalyst/app/transport/response"
-	"github.com/kosatnkn/catalyst/app/transport/response/transformers"
+	"github.com/kosatnkn/catalyst/app/http/request"
+	"github.com/kosatnkn/catalyst/app/http/request/unpackers"
+	"github.com/kosatnkn/catalyst/app/http/response"
+	"github.com/kosatnkn/catalyst/app/http/response/transformers"
 	"github.com/kosatnkn/catalyst/app/validator"
 	"github.com/kosatnkn/catalyst/domain/entities"
 	"github.com/kosatnkn/catalyst/domain/globals"
